Revert back to edge
Something about the v3.3 ruby installation is not right and we find
require errors on basic packages like io-console and bigdecimal in
downstream projects. This is likely why this is still using edge.

Down the line, the real fix will probably be to avoid this image
entirely and use the official alpine-based ruby images. That has been
deferred for now.